init: fix controlling tty for execve()
The forked processes used to set up a controlling tty. But this
did not properly work because they had to switch from a different
session than the parent's, leading to lots of output being lost
when the process exited and the session terminated. The solution
was to use 2 processes, one to perform the work, and the other
one to maintain the session the time needed for the chars to
reach the tty.