init: add the "br" command to branch without returning
Many scripts are written with a terminating "in /bin/sh", but this
causes some trouble because /bin/sh is passed init's args which are
invalid for sh.

A new "br" command (for "branch") works exactly like "ex" except
that it does not return in case of success. It simply does a pure
execve(). It also supports arguments if necessary.

So as of now, it is recommended that preinit scripts finish with
"br /bin/sh".