pcidev: add class-filtering support, small cleanups
It is now possible to limit the class of devices we want to look up.
By passing "-c 0x020000/0xffff00" for instance, we can focus on Ethernet
devices only.

Also, a nice improvement is that when listing only new modules (-n),
the code tries to avoid duplicate names when they are consecutive.