init: fix tty setting issues with Ctrl-C
The controlling TTY must be enabled only once we have forked the
new process, and it must be disabled with foreground restored to
init once the forked process dies. Since init cannot do that, we
have to use the intermediate process to do the job.

It apparently works correctly for all programs now, even when
background processes are still tied to the terminal.