init: fix terminal flushing after fork
Since the last terminal handling fixes, we lost the ability to
correctly drain output data. This means that running /bin/ps would
not have the time to display the output. This was caused by the
intermediate session.

Now we only have one session for the two processes (child and cleaner)
and the child runs in its own PGID. The cleaner only waits for the
terminal to drain output data and does not change its PGID at all.

This works correctly. Sub-processes correctly receive Ctrl-C, can
intercept it, and output data is correctly drained.