init: add support for devtmpfs at boot
When running from an initramfs, devtmpfs is not mounted by the kernel
and we have to do it ourselves. It is much more convenient when available
than a static /dev as it immediately presents all possible root devices
whatever their IDs.

We still create the missing entries according to the .preinit, and we
also set permissions based on those referenced there. That way the
.preinit serves as a reference for permissions even if major/minors
are different.