init: fix a major bug in interactive mode
A buffer overflow happens in interactive mode, where the prompt is copied
into a string as large as the level of braces and forgets to also allocate
enough room for the max string length. Moreover, the string copy is limited
to the max string length but this size was not allocated. The net effect is
init segfaulting in interactive mode when exiting a prompt using ".",
because the message displayed when entering the mode overwrites some context
info which his needed upon exit.