scripts/pcidev: support modules.alias as well
Starting with kernel 3.15, modules.pcimap is empty. Modules.alias
provides everything necessary even on older kernels, it's just that
the format is different. But "modprobe -R" can give us the module
name and is much faster at parsing the file than the previous loop.

So let's first try with modules.alias when it's present, and fall
back to modules.pcimap. This makes it compatible with newer kernels
and is faster to process the file.