init: cleanup: remove all the preset strings
These ones are hard to maintain and are confusing given that some are
declared as relative to the previous ones to get merged. Recent compilers
do a better job here and do not benefit from all of this anymore. By
replacing these references with the original text in the code, the
x86 code didn't change but the ARM code shrank byt 18 bytes. And
overall the code is much cleaner and easier to read.

Some strings are still dangerous when they don't start with a leading
"/" for certain paths, but it will be easier to review them now.