init: implement recursive mkdir()
This is the equivalent of mkdir -p. It now always ensures that the
intermediary directories are created and also takes care of stripping
the trailing slashes. The default "md" call has now been replaced and
this will be used for other situations where a sub-directory is needed.
Intermediary directories are created with mode 0755 when needed.