init: implement support for "tar {t,x,xv} archive.tar dir"
This is very convenient to extract modules as currently we're forced
to build and install busybox only to have tar. The implementation is
minimalist but sufficient. It supports uid/gid/mode, ignores mtime,
uname, or gname, supports the long file names having a prefix and
always extracts them relative to <dir>. It supports all standard unix
file types (chr,blk,fifo,file,dir,link,symlink,...). Note that the
implementation is not secure in that it doesn't protect against trying
to go over the directory by prepending many "../" but it's not needed
for initrds. Write errors are silently ignored.