init: implement "ls"
This "ls" command supports :
  -l to also print the file type at the beginning of the line
  -e to only report if a file exists in the directory in the status code

The "-" is optional, the option is considered when there are 2 args. The
directory is always the last one.

It's trivial to also report the file count, but we're missing an itoa()
function which is already implemented in nolibc though.

The whole function adds about 480 bytes to the executable, which is
reasonable compared to the benefits it provides, especially in places
where we used to require running a shell to check for file presence.