pause: introduce a new simple utility to maintain a live process
This one does nothing but wait for being killed without eating CPU,
it's simply a select() in a loop, the cheapest we can think of. If
called with -D it forks into background, closing all its FDs and
isolating itself into its own session.

The purpose mainly is to prevent a namespace from dying by always
leaving at least one process in it.