lcd: make lcdwrite conflict-resistant
It can happen during services startup that some services display info on
the LCD and that at the exact same time some async notifications are sent
(eg: VRRP switchover). This causes a conflict, one of the senders gets an
EBUSY response and aborts, leaving the previous message displayed.

Here instead we loop around a usleep() for no more than 1 second cumulated
to avoid this situation. Since LCD messages are short, it should not be an
issue.