scripts/hotplug: protect against potential spaces in DEVPATH
It's uncertain whether it's possible to get a device path having spaces
in its name/path but given the crap some SoC vendors can do in their BSPs,
better be safe. Also check that "loading" actually exists when stopping
it, to avoid a useless error when trying to write -1 there.