init: add a little description for each help command
This way it becomes easier to know how to use some of them. The help
string is made of a single long string with zeroes in the middle so
that we don't have to waste space using many pointers. This adds less
than 1 kB total for all commands, which probably is a reasonable
tradeoff.