flx: add --exclude-from to ignore certain files from a list during check
This will match the same option in tar and ease detection of certain
changes. Sadly, the code is impossible to make evolve and impossible
to follow, so the modification had to be performed in the output code
instead of the tree scan because it was not possible to figure how
to figure the paths nor what certain types correspond to :-/  At least
it works now.