init: make it possible to execute commands inline with -z
When the program is not called as PID 1, as "linuxrc" nor with a "rebuild"
argument, then the first argument will be checked for "-z". If it matches,
then the subsequent arguments will constitute a command to be executed inline.
For example, "init -z help" will show the help, and "init -z rd" will enter
interactive mode. This can be useful from scripts which need to benefit from
embedded commands, as well as to test new commands during development.