scripts/pcidev: deduplicate module names in the per-device list
For each device we build a list of candidate modules to be loaded, but
it sometimes happens that modprobe -q emits the same module multiple
times. Let's deduplicate this.

Note that this still doesn't deduplicate the modules list for loading
when multiple cards use the same driver, but this is normally achieved
by checking that it is already loaded, which is normally the case when
-m is used since it happens on the first invocation.