init: add a setenv command
The scripts rely quite a bit on environment variables but it is currently
not possible to create any due to the lack of allocator. The present patch
tries to address this by providing functions to measure the environment
size (bytes and number of entries) and to duplicate the environment into
a larger one that's storage-independent. This allows the caller to rely on
alloca() to store it in the calling function, which is OK when it's main().

Now thanks to this it becomes possible to set environment variables using
"se name value" and to clear them by passing no value. Note that an empty
value is properly added and differs from no value which is a deletion.