init: add "fp" to find the system partitions
The "fp" command will scan block devices in quest of the best
pair of system partitions. These ones must be consecutive and
of the exact same size on the same device, somewhere after an
ext* partition (/flash), and one of them must be of one of the
FS type passed in argument if any, or "squashfs" if not set.

When file system types are specified, they're in preference
order, and the pair of partitions matching the first fs types
will win over the last ones. If a single fs type (or none) is
set, then the first pair of partitions matching the condition
will be selected.

The goal is to use it like this:

  fp ${img} squashfs ext4 xfs ext2
  tn ${chosen_type}
  &mt ${chosen_part} /root ${chosen_type}
  |{
    tn ${second_type}
    &mt ${second_part} /root ${second_type}
  }
  &{
    mv /boot /root/boot
    cd /root
    mv . /
    sw .
    in /sbin/init
  }
  |{
    wk "Failed to find mountable system partition, rebooting in 3 seconds unless Enter pressed." 3
    |rb
    rd "Entering interactive debugging, type 'help' for supported commands."
  }

With "img" set to "active" or "image1" to select the first image,
or "backup" or "image2" to select the second one, and normally
passed via /proc/cmdline.

Such an initrd only requires very few entries (boot/, dev/, proc/,
root/, .preinit, init) and can be produced like this:

   mksquashfs boot dev proc root .preinit init ../preinit.ird

A cleaner example config file was added.