init: add a "tn" command to test for non-empty string
It's sometimes needed to be able to check for the existence of some
environment variables, and the easiest way for this is to let the
parser resolve them and test the resulting string, like in shell.
The "tn" command does this, with "tn ${foo}" returning true only if
foo is defined and not empty. This way it even allows to test the
variables learned from /proc/cmdline.