init: lp: fix existing mbr and partition detection
Issuing "lp" multiple times causes existing devices to be detected
multiple times. The reason is that the parent's duplication is not
detected by blkdev_find_parent(), and that partitions are systematically
appended instead of checking that they're already present. Let's change
this and don't blindly add a partition before checking if it was found.
Now issuing "lp" multiple times results in listing only one instance of
the same devices.