init: add kx to implement kexec_file_load()
This variant is extremely simple to use, the caller must only pass the
FDs of the kernel and initrd, and a pointer to the cmdline, then call
the syscall. Once the files are loaded, a reboot is triggered using
method CMD_KEXEC.

The method was introduced in kernel 2.6.27, so we rely on __NR_kexec_file_load
to detect its support and possibly define the missing LINUX_REBOOT_CMD_KEXEC
if needed. The code falls through the reboot code for simplicity and in order
to keep the code compact. There is a trick however, it was noticed that the
very first call often fails with EADDRNOTAVAIL, so we implement a retry
mechanism.

It was succesfully tested on arm64 with a 30MB kernel and 17MB initrd.

It's possible to ignore the initrd by passing "-". If no extra args are
passed, /proc/cmdline is used as the arguments. Otherwise if "-" is passed,
no args are used, otherwise the exact args are used.

A simple way to reboot using kexec is now:

   init -z kx kernel.img initrd.img