init: support type "auto" for "mt"
Since we can identify a number of file systems, let's make use of this to
simplify boot scripts by allowing type "auto" which triggers a detection
and uses the detected type. If the type is not detected, mount will fail.

At the moment, the properly detected block-device FS types are:

  vfat, ext2, ext3, ext4, xfs, squashfs3, squashfs

Other ones are (ubifs, jffs2, mbr, image types) are normally not usable
here.