init: add support for loading compressed modules
If a module doesn't start with "\x7F""ELF" then it's likely compressed,
so let's just pass MODULE_INIT_COMPRESSED_FILE in this case. It's only
supported since kernel 5.17 when CONFIG_MODULE_DECOMPRESS is set and
the proper module compression algorithm is supported, so it will simply
fail with other versions or configurations (as of 6.17, the kernel
supports a single decompression algorithm at once).