init: display a shortened errno name with errors
It's difficult to figure what happened when an error is reported, even
with the errno value. Let's add a table of the errno base (entries from
1 to 34) that are the most common. In order to save space, they're all
concatenated and delimited with a zero. The table is entirely scanned to
find the relevant code. This adds about 280 bytes total to the binary,
but for good.