init: do not replace console if already attached to a pipe
If any of stdin/stdout/stderr is connected to a pipe, then it means that
init was called under supervision of another process, which typically
matches what is done inside containers. In this case we do not want to
replace them with /dev/console. So let's add a test for this specific
case.